move: #1 解析器はドメインサービスが適切と判断し移動
